+++ src/ComplexSearch/icons/ClearIcon.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import SvgIcon from '../../icons/SvgIcon'
+import SvgIcon from '../../SvgIcon'
 
 export default function ClearIcon(props) {
   return (