+++ src/icons/forms/index.js
@@ -1,5 +1,4 @@
 export {default as TickIcon} from './TickIcon'
-export {default as Eye} from './Eye'
 export {default as EyeIcon} from './EyeIcon'
 export {default as ClosedEyeIcon} from './ClosedEyeIcon'
 export {default as QuestionIcon} from './QuestionIcon'