+++ src/index.js
@@ -1,7 +1,6 @@
-import * as icons from './icons'
-
-export {icons}
-
+export * from './icons/forms'
+export * from './icons/profiles'
+export * from './icons/services'
 export * from './theme'
 export * from './Toggle'
 export * from './Radio'
@@ -43,8 +42,8 @@ export * from './FormGroup'
 export {default as FieldStatus} from './FieldStatus'
 export * from './FieldStatus'
 
-export {default as SvgIcon} from './icons/SvgIcon'
-export * from './icons/SvgIcon'
+export {default as SvgIcon} from './SvgIcon'
+export * from './SvgIcon'
 
 export {default as Hint} from './Hint'
 export * from './Hint'